Endpoint with LaunchDarkly

+++ WeatherView/Models/PersonalWeatherViewModel.cs
@@ -1,5 +1,3 @@
-using System.Drawing;
-
 namespace WeatherView.Models
 {
     public class PersonalWeatherViewModel
@@ -7,7 +5,7 @@ namespace WeatherView.Models
         public string Name { get; set; }
         public string CityName { get; set; }
         public string CountryCode { get; set; }
-        public float Temperature { get; set; }
+        public double Temperature { get; set; }
         public string WeatherDescription { get; set; }
         public string Icon { get; set; }
         public string BackgroundColor { get; set; }
